services/orchestrator/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0
sqlalchemy==2.0.25
asyncpg==0.29.0
apscheduler==3.10.4
httpx==0.26.0
openai==1.12.0
json5==0.9.14
pytest==7.4.4
pytest-asyncio==0.23.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
